[front] chore: move server-side files from `lib/actions` to `lib/api/actions` (#22378)
