+++ biome-plugins/enforceClientTypesInPublicApi.grit
@@ -10,6 +10,7 @@ language js
 `"@dust-tt/client"` as $source where {
     $filename <: not r".*pages/api/v1/.*",
     $filename <: not r".*lib/actions/mcp_internal_actions/.*",
+    $filename <: not r".*lib/api/actions/servers/.*",
     $filename <: not r".*\.test\..*",
     $filename <: not r".*/connectors/.*",
     $filename <: not r".*/cli/.*",
@@ -22,7 +23,7 @@ language js
     $filename <: not r".*/core/.*",
     register_diagnostic(
         span = $source,
-        message = "Files outside /api/v1/ or lib/actions/mcp_internal_actions/ cannot import from '@dust-tt/client'.",
+        message = "Files outside /api/v1/, lib/actions/mcp_internal_actions/ or lib/api/actions/mcp_internal_actions/ cannot import from '@dust-tt/client'.",
         severity = "error"
     )
 }
