+++ front/lib/api/actions/servers/common/viz/instructions.ts
