+++ front/lib/api/actions/servers/microsoft_excel/tools/index.ts
@@ -1,11 +1,11 @@
 import { MCPError } from "@app/lib/actions/mcp_errors";
+import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import {
   getDriveItemEndpoint,
   getGraphClient,
   parseCellRef,
-} from "@app/lib/actions/mcp_internal_actions/servers/microsoft/utils";
-import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
-import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+} from "@app/lib/api/actions/servers/microsoft/utils";
 import { makeExcelRequest } from "@app/lib/api/actions/servers/microsoft_excel/helpers";
 import { MICROSOFT_EXCEL_TOOLS_METADATA } from "@app/lib/api/actions/servers/microsoft_excel/metadata";
 import { Err, Ok } from "@app/types/shared/result";
