+++ front/lib/api/actions/servers/run_agent/index.ts
@@ -4,13 +4,6 @@ import type {
   MCPProgressNotificationType,
   RunAgentQueryProgressOutput,
 } from "@app/lib/actions/mcp_internal_actions/output_schemas";
-import { getOrCreateConversation } from "@app/lib/actions/mcp_internal_actions/servers/run_agent/conversation";
-import { isTransientStreamError } from "@app/lib/actions/mcp_internal_actions/servers/run_agent/network_errors";
-import type {
-  ChildAgentBlob,
-  RunAgentBlockingEvent,
-} from "@app/lib/actions/mcp_internal_actions/servers/run_agent/types";
-import { makeToolBlockedAwaitingInputResponse } from "@app/lib/actions/mcp_internal_actions/servers/run_agent/types";
 import type {
   ToolDefinition,
   ToolHandlerExtra,
@@ -30,11 +23,18 @@ import {
   isServerSideMCPServerConfiguration,
 } from "@app/lib/actions/types/guards";
 import { RUN_AGENT_ACTION_NUM_RESULTS } from "@app/lib/actions/utils";
+import { getOrCreateConversation } from "@app/lib/api/actions/servers/run_agent/conversation";
 import {
   RUN_AGENT_CONFIGURABLE_PROPERTIES,
   RUN_AGENT_PLACEHOLDER_TOOL_NAME,
   RUN_AGENT_TOOL_SCHEMA,
 } from "@app/lib/api/actions/servers/run_agent/metadata";
+import { isTransientStreamError } from "@app/lib/api/actions/servers/run_agent/network_errors";
+import type {
+  ChildAgentBlob,
+  RunAgentBlockingEvent,
+} from "@app/lib/api/actions/servers/run_agent/types";
+import { makeToolBlockedAwaitingInputResponse } from "@app/lib/api/actions/servers/run_agent/types";
 import {
   getCitationsFromActions,
   getRefs,
