+++ front/lib/api/actions/servers/slack_personal/tools/index.ts
@@ -1,6 +1,14 @@
 import { getConnectionForMCPServer } from "@app/lib/actions/mcp_authentication";
 import { MCPError } from "@app/lib/actions/mcp_errors";
 import type { SearchResultResourceType } from "@app/lib/actions/mcp_internal_actions/output_schemas";
+import type {
+  ToolDefinition,
+  ToolHandlers,
+} from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { makePersonalAuthenticationError } from "@app/lib/actions/mcp_internal_actions/utils";
+import type { AgentLoopContextType } from "@app/lib/actions/types";
+import { SLACK_SEARCH_ACTION_NUM_RESULTS } from "@app/lib/actions/utils";
 import {
   executeListUserGroups,
   executePostMessage,
@@ -14,15 +22,7 @@ import {
   resolveChannelId,
   resolveUserDisplayName,
   SLACK_THREAD_LISTING_LIMIT,
-} from "@app/lib/actions/mcp_internal_actions/servers/slack/helpers";
-import type {
-  ToolDefinition,
-  ToolHandlers,
-} from "@app/lib/actions/mcp_internal_actions/tool_definition";
-import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
-import { makePersonalAuthenticationError } from "@app/lib/actions/mcp_internal_actions/utils";
-import type { AgentLoopContextType } from "@app/lib/actions/types";
-import { SLACK_SEARCH_ACTION_NUM_RESULTS } from "@app/lib/actions/utils";
+} from "@app/lib/api/actions/servers/slack/helpers";
 import { SLACK_PERSONAL_TOOLS_METADATA } from "@app/lib/api/actions/servers/slack_personal/metadata";
 import { getRefs } from "@app/lib/api/assistant/citations";
 import type { Authenticator } from "@app/lib/auth";
