+++ front/lib/api/actions/servers/sound_studio/tools/index.ts
@@ -1,10 +1,10 @@
 import { MCPError } from "@app/lib/actions/mcp_errors";
+import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import {
   getElevenLabsClient,
   streamToBase64,
-} from "@app/lib/actions/mcp_internal_actions/servers/elevenlabs/utils";
-import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
-import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+} from "@app/lib/api/actions/servers/elevenlabs/utils";
 import { SOUND_STUDIO_TOOLS_METADATA } from "@app/lib/api/actions/servers/sound_studio/metadata";
 import { Err, Ok } from "@app/types/shared/result";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
